requirements.txt
djangorestframework==3.15.1
python-dotenv==1.0.1
psycopg2-binary==2.9.9
drf-yasg==1.21.7
djangorestframework-csv==3.0.2